Exercice_Django/requirements-dev.txt
-r requirements.txt

# Testing
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3

# Type checking
mypy>=1.8
django-stubs>=4.2
djangorestframework-stubs>=3.14
types-python-dateutil